Few more cleanups

+++ Engine/AssetID.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "ResourceType.h"
 
-using AssetID = uint32_t;
+using AssetID = uint64_t;
 
 static constexpr AssetID NULL_ASSET = (~static_cast <AssetID>(0)); // Can collide though...
